+++ src/groups/power_group.jl
@@ -309,6 +309,11 @@ function ManifoldsBase.log!(
     end
     return X
 end
+function ManifoldsBase.log!(
+    PoG::LieGroup{𝔽,Op,M}, X, ::Identity{Op}, ::Identity{Op}
+) where {𝔽,Op<:PowerGroupOperation,M<:ManifoldsBase.AbstractPowerManifold}
+    return zero_vector!(LieAlgebra(PoG), X)
+end
 
 function Base.show(
     io::IO, G::LieGroup{𝔽,Op,M}