+++ src/interface.jl
@@ -11,7 +11,7 @@ on elements of a Lie group ``$(_math(:G))``.
 abstract type AbstractGroupOperation end
 
 """
-    LieAlgebraOrthogonalBasis{𝔽} <: ManifoldsBase.AbstractOrthogonalBasis{𝔽,ManifoldsBase.TangentSpaceType}
+    DefaultLieAlgebraOrthogonalBasis{𝔽} <: ManifoldsBase.AbstractOrthogonalBasis{𝔽,ManifoldsBase.TangentSpaceType}
 
 Specify an orthogonal basis for a Lie algebra.
 This is used as the default within [`hat`](@ref) and [`vee`](@ref).
@@ -23,10 +23,10 @@ at the [`identity_element`](@ref) on the [`base_manifold](@ref base_manifold(::L
     In order to implement the corresponding [`get_coordinates`](@ref) and [`get_vector`](@ref) functions,
     define `get_coordiinates_lie(::LieGroup, p, X, N)` and `get_vector_lie(::LieGroup, p, X, N)`, resp.
 """
-struct LieAlgebraOrthogonalBasis{𝔽} <:
+struct DefaultLieAlgebraOrthogonalBasis{𝔽} <:
        ManifoldsBase.AbstractOrthogonalBasis{𝔽,ManifoldsBase.TangentSpaceType} end
-function LieAlgebraOrthogonalBasis(𝔽::ManifoldsBase.AbstractNumbers=ℝ)
-    return LieAlgebraOrthogonalBasis{𝔽}()
+function DefaultLieAlgebraOrthogonalBasis(𝔽::ManifoldsBase.AbstractNumbers=ℝ)
+    return DefaultLieAlgebraOrthogonalBasis{𝔽}()
 end
 
 """
@@ -90,20 +90,22 @@ Identity(::Type{O}) where {O<:AbstractGroupOperation} = Identity{O}()
 # Internal pass through for coordinates and vectors
 
 @inline function ManifoldsBase._get_coordinates(
-    G::LieGroup, p, X, B::LieAlgebraOrthogonalBasis
+    G::LieGroup, p, X, B::DefaultLieAlgebraOrthogonalBasis
 )
     return get_coordinates_lie(G, p, X, number_system(B))
 end
 @inline function ManifoldsBase._get_coordinates!(
-    G::LieGroup, Y, p, X, B::LieAlgebraOrthogonalBasis
+    G::LieGroup, Y, p, X, B::DefaultLieAlgebraOrthogonalBasis
 )
     return get_coordinates_lie!(G, Y, p, X, number_system(B))
 end
-@inline function ManifoldsBase._get_vector(G::LieGroup, p, c, B::LieAlgebraOrthogonalBasis)
+@inline function ManifoldsBase._get_vector(
+    G::LieGroup, p, c, B::DefaultLieAlgebraOrthogonalBasis
+)
     return get_vector_lie(G, p, c, number_system(B))
 end
 @inline function ManifoldsBase._get_vector!(
-    G::LieGroup, Y, p, c, B::LieAlgebraOrthogonalBasis
+    G::LieGroup, Y, p, c, B::DefaultLieAlgebraOrthogonalBasis
 )
     return get_vector_lie!(G, Y, p, c, number_system(B))
 end
@@ -443,13 +445,15 @@ ManifoldsBase.get_coordinates(G::LieGroup, g, X, B::ManifoldsBase.AbstractBasis)
 ManifoldsBase.get_coordinates!(G::LieGroup, c, g, X, B::ManifoldsBase.AbstractBasis)
 
 function get_coordinates_lie(G::LieGroup, g, X, N)
-    return get_coordinates(
-        base_manifold(G), identity_element(G), X, ManifoldsBase.DefaultOrthogonalBasis(N)
+    c = allocate_result(
+        G, get_coordinates, identity_element(G), X, DefaultLieAlgebraOrthogonalBasis(N)
     )
+
+    return get_coordinates_lie!(G, c, g, X, N)
 end
-function get_coordinates_lie!(G::LieGroup, Y, g, X, N)
+function get_coordinates_lie!(G::LieGroup, c, g, X, N)
     return get_coordinates!(
-        base_manifold(G), Y, identity_element(G), X, ManifoldsBase.DefaultOrthogonalBasis(N)
+        base_manifold(G), c, identity_element(G), X, ManifoldsBase.DefaultOrthogonalBasis(N)
     )
 end
 
@@ -481,9 +485,8 @@ ManifoldsBase.get_vector(G::LieGroup, g, c, B::ManifoldsBase.AbstractBasis)
 ManifoldsBase.get_vector!(G::LieGroup, X, g, c, B::ManifoldsBase.AbstractBasis)
 
 @inline function get_vector_lie(G::LieGroup, g, c, N)
-    return get_vector(
-        base_manifold(G), identity_element(G), c, ManifoldsBase.DefaultOrthogonalBasis(N)
-    )
+    X = zero_vector(G, Identity(G))
+    return get_vector_lie!(G, X, g, c, N)
 end
 @inline function get_vector_lie!(G::LieGroup, Y, g, c, N)
     return get_vector!(
@@ -509,7 +512,7 @@ The computation can be performed in-place of `X`.
 The inverse of `hat` is [`vee`](@ref).
 
 Technically, `hat` is a specific case of [`get_vector`](@ref) and is implemented using the
-[`LieAlgebraOrthogonalBasis`](@ref)
+[`DefaultLieAlgebraOrthogonalBasis`](@ref)
 """
 
 # function hat end
@@ -714,8 +717,8 @@ function ManifoldsBase.isapprox(
 end
 
 _doc_jacobian_conjugate = """
-    jacobian_conjugate(G::LieGroup, g, h, B::ManifoldsBase.AbstractBasis=LieAlgebraOrthogonalBasis())
-    jacobian_conjugate!(G::LieGroup, J, g, h, B::ManifoldsBase.AbstractBasis=LieAlgebraOrthogonalBasis())
+    jacobian_conjugate(G::LieGroup, g, h, B::ManifoldsBase.AbstractBasis=DefaultLieAlgebraOrthogonalBasis())
+    jacobian_conjugate!(G::LieGroup, J, g, h, B::ManifoldsBase.AbstractBasis=DefaultLieAlgebraOrthogonalBasis())
 
 Compute the Jacobian of the [`conjugate`](@ref) ``c_g(h) = g$(_math(:∘))h$(_math(:∘))g^{-1}``,
 with respect to an [`AbstractBasis`](@extref `ManifoldsBase.AbstractBasis`).
@@ -727,11 +730,11 @@ with respect to the given basis.
     For the case that `h` is the [`Identity`](@ref) and the relation of ``D(c_g(h))[X]``
     to the [`adjoint`](@ref) ``$(_math(:Ad))(g)``, the Jacobian then sometimes called “adjoint matrix”,
     e.g. in [SolaDerayAtchuthan:2021](@cite), when choosing as a basis the
-    [`LieAlgebraOrthogonalBasis`](@ref)`()` that is used for [`hat`](@ref) and [`vee`](@ref).
+    [`DefaultLieAlgebraOrthogonalBasis`](@ref)`()` that is used for [`hat`](@ref) and [`vee`](@ref).
 """
 @doc "$(_doc_jacobian_conjugate)"
 function jacobian_conjugate(
-    G::LieGroup, g, h, B::ManifoldsBase.AbstractBasis=LieAlgebraOrthogonalBasis()
+    G::LieGroup, g, h, B::ManifoldsBase.AbstractBasis=DefaultLieAlgebraOrthogonalBasis()
 )
     J = ManifoldsBase.allocate_result(G, jacobian_conjugate, g, h, B)
     return jacobian_conjugate!(G, J, g, h, B)
@@ -740,7 +743,7 @@ end
 function jacobian_conjugate! end
 @doc "$(_doc_jacobian_conjugate)"
 jacobian_conjugate!(
-    ::LieGroup, J, g, h; B::ManifoldsBase.AbstractBasis=LieAlgebraOrthogonalBasis()
+    ::LieGroup, J, g, h; B::ManifoldsBase.AbstractBasis=DefaultLieAlgebraOrthogonalBasis()
 )
 
 _doc_log = """
@@ -772,11 +775,22 @@ function ManifoldsBase.log(G::LieGroup, g, h)
     log!(G, X, g, h)
     return X
 end
+function ManifoldsBase.log(
+    G::LieGroup{𝔽,O}, ::Identity{O}, ::Identity{O}
+) where {𝔽,O<:AbstractGroupOperation}
+    return zero_vector(LieAlgebra(G))
+end
 
 @doc "$_doc_log"
 function ManifoldsBase.log!(G::LieGroup, X, g, h)
     log!(G, X, Identity(G), compose(G, inv(G, g), h))
     return h
+end
+function ManifoldsBase.log!(
+    ::LieGroup{𝔽,O}, X, ::Identity{O}, ::Identity{O}
+) where {𝔽,O<:AbstractGroupOperation}
+    zero_vector!(LieAlgebra(G), X)
+    return X
 end
 
 _doc_log_id = """
@@ -856,7 +870,7 @@ _doc_vee = """
     vee!(G::LieGroup, c, X)
 
 Compute the vee map ``(⋅)^∨`` that maps a tangent vector `X` from the [`LieAlgebra`](@ref)
-to its coordinates with respect to the [`LieAlgebraOrthogonalBasis`](@ref) basis in the Lie algebra
+to its coordinates with respect to the [`DefaultLieAlgebraOrthogonalBasis`](@ref) basis in the Lie algebra
 
 ```math
 X = $(_tex(:sum))_{i∈$(_tex(:Cal,"I"))} c_iB_i,
@@ -869,7 +883,7 @@ The computation can be performed in-place of `c`.
 The inverse of `vee` is [`hat`](@ref).
 
 Technically, `vee` is a specific case of [`get_coordinates`](@ref) and is implemented using
-the [`LieAlgebraOrthogonalBasis`](@ref).
+the [`DefaultLieAlgebraOrthogonalBasis`](@ref).
 """
 
 # function vee end